third_party/WebKit/LayoutTests/fast/workers/chromium/worker-crash-with-invalid-location-expected.txt

Blocked access to external URL http://example.com/worker.js
CONSOLE ERROR: line 18: Failed to start loading http://example.com/worker.js
Test worker fetch of blocked url. Should print a "PASS" statement.

PASS
